refactor: 컴포넌트 폴더 구조 변경 (페이지 단위에서 productList 분리)
